[pug-lang] added `ClassEnv`, moved `tiExpr` and `tiProgram` to Infer.{c|h}
