+++ example/pug-lang/src/types/TypeInfer.c
@@ -2,11 +2,13 @@
 
 #include "types/TypeInfer.h"
 
-static inline Infered(Type) InferedType(List(Pred) ps, Type t) {
+// -----------------------------------------------------------------------
+Infered(Type) InferedType(List(Pred) ps, Type t) {
   return (Infered(Type)){.ps = ps, .t = t};
 }
 
-static List(Pred) appendPreds(List(Pred) ps1, List(Pred) ps2) {
+// -----------------------------------------------------------------------
+List(Pred) appendPreds(List(Pred) ps1, List(Pred) ps2) {
   if (!ps1) {
     return ps2;
   }
@@ -86,356 +88,5 @@ action(freshInst, Scheme, Qual(Type)) {
     }
     Instantiate(Qual(Type)) I = trait(Instantiate(Qual(Type)));
     A_RETURN(I.inst(ts, sc.qt));
-  }
-}
-
-// -----------------------------------------------------------------------
-action(ti_label0, TI(Infered(Type)), Expr, Infered(Type)) {
-  A_DO_WITH(ti, e) {
-    A_RUN(getState(), s);
-    TIResult(Infered(Type)) r = runTI(ti, s);
-    A_RUN(putState(r.state));
-    if (r.success) {
-      A_RETURN(r.ok);
-    }
-    CharBuff b = {0};
-    mem_printf(&b, "%s\n | %s", r.err.msg, trait(Show(Expr)).show(e));
-    r.err.msg = b.data;
-    A_FAIL(r.err);
-  }
-}
-
-TI(Infered(Type)) ti_label(TI(Infered(Type)) ti, Expr e) {
-  switch (e->id) {
-  case SEQ:
-  case BLK:
-    return ti;
-  default:
-    return ti_label0(ti, e);
-  }
-}
-
-// -----------------------------------------------------------------------
-action(tiExprVar, List(Assump), Expr, Infered(Type)) {
-  A_DO_WITH(as, e) {
-    Maybe(Scheme) sc = t_find(e->ident, as);
-    if (sc.none) {
-      CharBuff b = {0};
-      mem_printf(&b, "Undefined variable - %s", e->ident);
-      A_FAIL((TypeError){b.data});
-    }
-    A_RUN(freshInst(sc.value), qt);
-    A_RETURN(InferedType(qt.ps, qt.t));
-  }
-}
-
-action(tiExprLambda, List(Assump), Expr, Infered(Type)) {
-  A_DO_WITH(as, e) {
-    TypeT T = trait(Type);
-    A_RUN(newTVar(trait(Kind).Star()), a);
-    A_RUN(newTVar(trait(Kind).Star()), b);
-    A_RUN(newTVar(trait(Kind).Star()), f);
-    A_RUN(unify(T.func(a, b), f));
-    Scheme sc = toScheme(a); // TODO !?
-    as = t_add(e->lhs->ident, sc, as);
-    A_RUN(tiExpr(as, e->rhs), c);
-    A_RUN(unify(c.t, b));
-    A_RETURN(InferedType(NULL, f));
-  }
-}
-
-action(tiExprApply, List(Assump), Expr, Infered(Type)) {
-  A_DO_WITH(as, e) {
-    TypeT T = trait(Type);
-    A_RUN(tiExpr(as, e->lhs), a);
-    A_RUN(tiExpr(as, e->rhs), b);
-    A_RUN(newTVar(trait(Kind).Star()), t);
-    A_RUN(unify(T.func(b.t, t), a.t));
-    List(Pred) ps = appendPreds(a.ps, b.ps);
-    A_RETURN(InferedType(ps, t));
-  }
-}
-
-action(tiExprIfelse, List(Assump), Expr, Infered(Type)) {
-  A_DO_WITH(as, e) {
-    Type Bool = trait(Type).tcon_bool();
-    A_RUN(tiExpr(as, e->lhs), c);
-    A_RUN(unify(c.t, Bool));
-    A_RUN(tiExpr(as, e->rhs->lhs), a);
-    A_RUN(tiExpr(as, e->rhs->rhs), b);
-    A_RUN(unify(a.t, b.t));
-    List(Pred) ps = appendPreds(a.ps, b.ps);
-    A_RETURN(InferedType(ps, a.t));
-  }
-}
-
-action(tiExprBlk, List(Assump), Expr, Infered(Type)) {
-  A_DO_WITH(as, e) {
-    A_RUN(tiExpr(as, e->rhs), x);
-    A_RETURN(x);
-  }
-}
-
-action(tiExprSeq, List(Assump), Expr, Infered(Type)) {
-  A_DO_WITH(as, e) {
-    TypeT T = trait(Type);
-    switch (e->lhs->id) {
-    case DECLVAR: {
-      Scheme sc = t_gen(as, e->lhs->rhs->texpr);
-      as = t_add(e->lhs->lhs->ident, sc, as);
-      A_RUN(tiExpr(as, e->rhs), x);
-      A_RETURN(x);
-    }
-    case LET: {
-      Maybe(Scheme) sc = t_find(e->lhs->lhs->ident, as);
-      if (!sc.none) {
-        A_RUN(tiExpr(as, e->lhs->rhs), a);
-        A_RUN(getSubst(), sub);
-        Scheme sc = t_gen(as, t_apply_subst(sub, a.t));
-        as = t_add(e->lhs->lhs->ident, sc, as);
-        A_RUN(tiExpr(as, e->rhs), x);
-        List(Pred) ps = appendPreds(a.ps, x.ps);
-        A_RETURN(InferedType(ps, x.t));
-      } else {
-        if (e->lhs->rhs->id == LAMBDA) {
-          A_RUN(newTVar(trait(Kind).Star()), a);
-          A_RUN(newTVar(trait(Kind).Star()), b);
-          A_RUN(newTVar(trait(Kind).Star()), f);
-          A_RUN(unify(T.func(a, b), f));
-          Scheme sc = toScheme(f); // TODO !?
-          as = t_add(e->lhs->lhs->ident, sc, as);
-        }
-        A_RUN(tiExpr(as, e->lhs->rhs), a);
-        A_RUN(getSubst(), sub);
-        Scheme sc = t_gen(as, t_apply_subst(sub, a.t));
-        as = t_add(e->lhs->lhs->ident, sc, as);
-        A_RUN(tiExpr(as, e->rhs), x);
-        A_RETURN(x);
-      }
-    }
-    default: {
-      A_RUN(tiExpr(as, e->lhs));
-      A_RUN(tiExpr(as, e->rhs), t);
-      A_RETURN(t);
-    }
-    }
-  }
-}
-
-action(tiExprDeclvar, List(Assump), Expr, Infered(Type)) {
-  A_DO_WITH(as, e) {
-    // TODO what should we do here?
-    A_RETURN(InferedType(NULL, e->rhs->texpr));
-  }
-}
-
-action(tiExprAssign, List(Assump), Expr, Infered(Type)) {
-  A_DO_WITH(as, e) {
-    A_RUN(tiExpr(as, e->lhs), a);
-    A_RUN(tiExpr(as, e->rhs), b);
-    A_RUN(unify(a.t, b.t));
-    List(Pred) ps = appendPreds(a.ps, b.ps);
-    A_RETURN(InferedType(ps, a.t));
-  }
-}
-
-action(tiExprLet, List(Assump), Expr, Infered(Type)) {
-  A_DO_WITH(as, e) {
-    Maybe(Scheme) sc = t_find(e->lhs->ident, as);
-    if (!sc.none) {
-      A_RUN(freshInst(sc.value), qt);
-      A_RUN(tiExpr(as, e->rhs), a);
-      A_RUN(unify(qt.t, a.t));
-      List(Pred) ps = appendPreds(qt.ps, a.ps);
-      A_RETURN(InferedType(ps, a.t));
-    } else {
-      if (e->rhs->id == LAMBDA) {
-        TypeT T = trait(Type);
-        A_RUN(newTVar(trait(Kind).Star()), a);
-        A_RUN(newTVar(trait(Kind).Star()), b);
-        A_RUN(newTVar(trait(Kind).Star()), f);
-        A_RUN(unify(T.func(a, b), f));
-        Scheme sc = toScheme(f); // TODO !?
-        as = t_add(e->lhs->ident, sc, as);
-      }
-      A_RUN(tiExpr(as, e->rhs), x);
-      A_RETURN(x);
-    }
-  }
-}
-
-action(tiExprOrAnd, List(Assump), Expr, Infered(Type)) {
-  A_DO_WITH(as, e) {
-    Type Bool = trait(Type).tcon_bool();
-    A_RUN(tiExpr(as, e->lhs), a);
-    A_RUN(tiExpr(as, e->rhs), b);
-    A_RUN(unify(a.t, b.t));
-    A_RUN(unify(Bool, b.t));
-    A_RUN(newTVar(trait(Kind).Star()), t);
-    A_RUN(unify(Bool, t));
-    A_RETURN(InferedType(NULL, t));
-  }
-}
-
-action(tiExprComparisson, List(Assump), Expr, Infered(Type)) {
-  A_DO_WITH(as, e) {
-    Type Bool = trait(Type).tcon_bool();
-    A_RUN(tiExpr(as, e->lhs), a);
-    A_RUN(tiExpr(as, e->rhs), b);
-    A_RUN(unify(a.t, b.t));
-    A_RUN(newTVar(trait(Kind).Star()), t);
-    A_RUN(unify(Bool, t));
-    A_RETURN(InferedType(NULL, t));
-  }
-}
-
-action(tiExprArithmetic, List(Assump), Expr, Infered(Type)) {
-  A_DO_WITH(as, e) {
-    Type Int = trait(Type).tcon_int();
-    A_RUN(tiExpr(as, e->lhs), a);
-    A_RUN(tiExpr(as, e->rhs), b);
-    A_RUN(unify(a.t, b.t));
-    A_RUN(unify(Int, b.t));
-    A_RUN(newTVar(trait(Kind).Star()), t);
-    A_RUN(unify(Int, t));
-    A_RETURN(InferedType(NULL, t));
-  }
-}
-
-action(tiExprNeg, List(Assump), Expr, Infered(Type)) {
-  A_DO_WITH(as, e) {
-    Type Int = trait(Type).tcon_int();
-    A_RUN(tiExpr(as, e->rhs), t);
-    A_RUN(unify(Int, t.t));
-    A_RETURN(t);
-  }
-}
-
-action(tiExprNot, List(Assump), Expr, Infered(Type)) {
-  A_DO_WITH(as, e) {
-    Type Bool = trait(Type).tcon_bool();
-    Type Int = trait(Type).tcon_int();
-    A_RUN(tiExpr(as, e->rhs), a);
-    A_RUN(getSubst(), sub);
-    Type b = t_apply_subst(sub, a.t);
-    Eq(Type) E = trait(Eq(Type));
-    if (E.eq(b, Int) || E.eq(b, Bool)) {
-      A_RETURN(a);
-    }
-    A_FAIL((TypeError){"Type mismatch"});
-  }
-}
-
-action(tiExprPrint, List(Assump), Expr, Infered(Type)) {
-  A_DO_WITH(as, e) {
-    Type Unit = trait(Type).tcon_unit();
-    A_RUN(tiExpr(as, e->rhs));
-    A_RETURN(InferedType(NULL, Unit));
-  }
-}
-
-action(tiExprCon, List(Assump), Expr, Infered(Type)) {
-  A_DO_WITH(as, e) {
-    ExprT E = trait(Expr);
-    A_RUN(tiExpr(as, E.var(e->ident)), t);
-    A_RETURN(t);
-  }
-}
-
-action(tiExprCapply, List(Assump), Expr, Infered(Type)) {
-  A_DO_WITH(as, e) {
-    ExprT E = trait(Expr);
-    A_RUN(tiExpr(as, E.apply(e->lhs, e->rhs)), t);
-    A_RETURN(t);
-  }
-}
-
-action(tiExprFail, List(Assump), Expr, Infered(Type)) {
-  A_DO_WITH(as, e) {
-    A_FAIL((TypeError){"Invalid expr"});
-  }
-}
-
-action(tiLiteral, List(Assump), Literal, Infered(Type)) {
-  A_DO_WITH(as, lit) {
-    TypeT T = trait(Type);
-    switch (lit.id) {
-    case LIT_INTEGER:
-      A_RETURN(InferedType(NULL, T.tcon_int()));
-    case LIT_UNIT:
-      A_RETURN(InferedType(NULL, T.tcon_unit()));
-    case LIT_TRUE:
-    case LIT_FALSE:
-      A_RETURN(InferedType(NULL, T.tcon_bool()));
-    default:
-      A_FAIL((TypeError){"Illegal Literal"});
-    }
-  }
-}
-
-static ACTION(Infered(Type)) tiExpr0(List(Assump) as, Expr e) {
-  switch (e->id) {
-  case VAR:
-    return tiExprVar(as, e);
-  case LAMBDA:
-    return tiExprLambda(as, e);
-  case APPLY:
-    return tiExprApply(as, e);
-  case IFELSE:
-    return tiExprIfelse(as, e);
-  case BLK:
-    return tiExprBlk(as, e);
-  case SEQ:
-    return tiExprSeq(as, e);
-  case DECLVAR:
-    return tiExprDeclvar(as, e);
-  case ASSIGN:
-    return tiExprAssign(as, e);
-  case LET:
-    return tiExprLet(as, e);
-  case OR:
-  case AND:
-    return tiExprOrAnd(as, e);
-  case EQ:
-  case NEQ:
-  case LE:
-  case LT:
-  case GT:
-  case GE:
-    return tiExprComparisson(as, e);
-  case ADD:
-  case SUB:
-  case MUL:
-  case DIV:
-  case MOD:
-    return tiExprArithmetic(as, e);
-  case NEG:
-    return tiExprNeg(as, e);
-  case NOT:
-    return tiExprNot(as, e);
-  case LITERAL:
-    return tiLiteral(as, e->literal);
-  case PRINT:
-    return tiExprPrint(as, e);
-  case CON:
-    return tiExprCon(as, e);
-  case CAPPLY:
-    return tiExprCapply(as, e);
-  default:
-    return tiExprFail(as, e);
-  }
-}
-
-TI(Infered(Type)) tiExpr(List(Assump) as, Expr e) {
-  return ti_label(tiExpr0(as, e), e);
-}
-
-// -----------------------------------------------------------------------
-action(tiProgram, List(Assump), Expr, Infered(Type)) {
-  A_DO_WITH(as, e) {
-    A_RUN(tiExpr(as, e), a);
-    A_RUN(getSubst(), sub);
-    a.t = t_apply_subst(sub, a.t);
-    A_RETURN(a);
   }
 }
